refactor: Reorganize files according to clean architecture
