+++ app/src/androidTest/java/campus/tech/kakao/map/MapActivityUITest.kt
@@ -8,8 +8,8 @@ import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.rule.ActivityTestRule
-import campus.tech.kakao.map.presentation.MapActivity
-import campus.tech.kakao.map.presentation.PlaceActivity
+import campus.tech.kakao.map.ui.MapActivity
+import campus.tech.kakao.map.ui.PlaceActivity
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
