+++ app/src/androidTest/java/campus/tech/kakao/map/PlaceActivityUnitTest.kt
@@ -6,9 +6,9 @@ import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import androidx.test.core.app.ActivityScenario
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import campus.tech.kakao.map.data.PlaceDataModel
+import campus.tech.kakao.map.domain.PlaceDataModel
 import campus.tech.kakao.map.data.PlaceDatabaseAccess
-import campus.tech.kakao.map.presentation.PlaceActivity
+import campus.tech.kakao.map.ui.PlaceActivity
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
