+++ app/src/main/java/campus/tech/kakao/map/data/PlaceRepository.kt
@@ -2,8 +2,7 @@ package campus.tech.kakao.map.data
 
 import android.util.Log
 import campus.tech.kakao.map.BuildConfig
-import campus.tech.kakao.map.domain.RetrofitInstance
-import campus.tech.kakao.map.domain.RetrofitLocalKeywordService
+import campus.tech.kakao.map.domain.PlaceDataModel
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
