+++ app/src/main/java/campus/tech/kakao/map/data/RetrofitLocalCategoryService.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map.domain
+package campus.tech.kakao.map.data
 
 import campus.tech.kakao.map.data.SearchResult
 import retrofit2.Call
@@ -6,10 +6,10 @@ import retrofit2.http.GET
 import retrofit2.http.Header
 import retrofit2.http.Query
 
-interface RetrofitLocalKeywordService {
-    @GET("/v2/local/search/keyword.json")
-    fun searchPlaceByKeyword(
+interface RetrofitLocalCategoryService {
+    @GET("/v2/local/search/category.json")
+    fun searchPlaceByCategory(
         @Header("Authorization") key: String,
-        @Query("query") query: String
+        @Query("category_group_code") categoryGroupCode: String
     ): Call<SearchResult>
 }
