+++ app/src/main/java/campus/tech/kakao/map/domain/PlaceDataModel.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map.data
+package campus.tech.kakao.map.domain
 
 data class PlaceDataModel(
     val name: String,
