+++ app/src/main/java/campus/tech/kakao/map/ui/KakaoMapApplication.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map.presentation
+package campus.tech.kakao.map.ui
 
 import android.app.Application
 import campus.tech.kakao.map.BuildConfig
