+++ app/src/main/res/layout/map_layout.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.MapActivity">
+    tools:context=".ui.MapActivity">
 
 
     <com.kakao.vectormap.MapView
